+++ includes/class.DpsPxPayWpscResponseRequest.php
@@ -36,7 +36,7 @@ class DpsPxPayWpscResponseRequest extends DpsPxPayWpscResponse {
 		parent::loadResponse($response);
 
 		if (empty($this->URI) && (!empty($this->Reco) || !empty($this->ResponseText))) {
-			$errors = array();
+			$errors = [];
 			if (!empty($this->ResponseText)) {
 				$errors[] = $this->ResponseText;
 			}