+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Changelog
 
+### 1.1.0, 2018-11-22
+
+* changed: requires minimum PHP version 5.6 (recommend version 7.2 or greater)
+* tested to WordPress 5.0
+
 ### 1.0.1, 2016-11-18
 
 * tested to WordPress 4.7